[native] Move writer test to a central place
